+++ R/wrf_put.R
@@ -1,8 +1,8 @@
-#' Funtion to write variables in emission files
+#' Function to write variables in emission files
 #'
 #' @description Extract variable
 #'
-#' @param file name of file interactively (defoult) or specified
+#' @param file name of file interactively (default) or specified
 #' @param name name of the variable (any variable)
 #' @param POL input
 #'